83 ajouter la progression pour les fichiers states

+++ App.Core/Models/StateManagerModel.cs
@@ -7,9 +7,9 @@
         public string SourceFilePath { get; set; } = string.Empty;                          // Path of the source file
         public string TargetFilePath { get; set; } = string.Empty;                          // Path of the target file
         public string State { get; set; } = "END";                                          // State of the save
-        public int TotalFilesToCopy { get; set; } = 0;                                      // Total files to copy
+        public long TotalFilesToCopy { get; set; } = 0;                                      // Total files to copy
         public long TotalFilesSize { get; set; } = 0;                                       // Total size of the files to copy
-        public int NbFilesLeftToDo { get; set; } = 0;                                       // Number of files left to copy
+        public long NbFilesLeftToDo { get; set; } = 0;                                       // Number of files left to copy
         public float Progression { get; set; } = 0;                                         // Progression of the save
 
     }
